% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.telemetry-package.R
\docType{package}
\name{impact.telemetry-package}
\alias{impact.telemetry}
\alias{impact.telemetry-package}
\title{impact.telemetry: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Iyed GHEDAMSI \email{iyed.ghedamsi@impact-initiatives.org}

Other contributors:
\itemize{
  \item IMPACT Initiatives [copyright holder]
}

}
\keyword{internal}
