% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact.telemetry-package.R
\docType{package}
\name{impact.telemetry-package}
\alias{impact.telemetry}
\alias{impact.telemetry-package}
\title{impact.telemetry: A Function Telemetry Toolkit for IMPACT Initiatives}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolkit for adding telemetry to R functions, intended for use within the IMPACT Initiatives ecosystem. Utilizes a wrapper function to log execution metadata to a file. The configuration and data schema are tailored for IMPACT's internal package monitoring requirements.
}
\author{
\strong{Maintainer}: Iyed GHEDAMSI \email{iyed.ghedamsi@impact-initiatives.org}

Other contributors:
\itemize{
  \item IMPACT Initiatives [copyright holder]
}

}
\keyword{internal}
